Fix broadcast

+++ server/src/protocol/broadcast_utils.c
@@ -95,7 +95,7 @@ int get_sound_player(player_t *p1, player_t *p2)
 	pl2 = correct_orientation_player(p2);
 	x = (int) pl1->x - (int) pl2->x;
 	y = (int) pl1->y - (int)pl2->y;
-	loop_segment(&x, &y, (server.map).width, (server.map).height);
+	loop_segment(&x, &y, (int) (server.map).width, (int) (server.map).height);
 	valid =
 	get_valid_broadcast_direction(
 		get_broadcast_orientation(x, y, pl2->orientation));
